Move validation util tests to one location
